Update SIP status values (refs #1160)
- Set SIP status to "failed" when the ingest workflow fails
- Update the internal SIP status enum with the new values
- Update the database SIP status column to an enum type with the new
  status values
- Regenerate the Ent, API (Goa) and dashboard SIP status values
- Update the Dashboard SIPs status tabs
- Update the SIPs legend with the new statuses and descriptions

+++ internal/persistence/ent/client/sip.go
@@ -25,7 +25,7 @@ func (c *client) CreateSIP(ctx context.Context, s *datatypes.SIP) error {
 
 	q := c.ent.SIP.Create().
 		SetName(s.Name).
-		SetStatus(int8(s.Status)) // #nosec G115 -- constrained value.
+		SetStatus(s.Status)
 
 	// Add optional fields.
 	if s.AIPID.Valid {
@@ -81,7 +81,7 @@ func (c *client) UpdateSIP(
 	// Set required column values.
 	q := tx.SIP.UpdateOneID(id).
 		SetName(up.Name).
-		SetStatus(int8(up.Status)) // #nosec G115 -- constrained value.
+		SetStatus(up.Status)
 
 	// Set nullable column values.
 	if up.AIPID.Valid {
